backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.15.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
